+++ dzVents/runtime/Utils.lua
@@ -7,7 +7,7 @@ local self = {
 	LOG_MODULE_EXEC_INFO = 2,
 	LOG_INFO = 3,
 	LOG_DEBUG = 4,
-	DZVERSION = '2.4.20',
+	DZVERSION = '2.4.21',
 }
 
 function self.fileExists(name)
@@ -27,6 +27,16 @@ function self.stringSplit(text, sep)
 		table.insert(t, str)
 	end
 	return t
+end
+
+function self.inTable(searchTable, element)
+	local res = res
+	for k, v in pairs(searchTable) do
+		if type(v) == 'table' then res = self.inTable(v, element) end
+		res = res or (( tostring(k) == tostring(element) and 'key' ) or ( tostring(v) == tostring(element) and 'value' ))
+		if res then return res end
+	end
+	return false
 end
 
 function self.osExecute(cmd)
@@ -55,6 +65,7 @@ function self.urlEncode(str, strSub)
 	end
 	return str
 end
+
 function self.urlDecode(str, strSub)
 
 	local hex2Char = function(x)
@@ -183,7 +194,7 @@ function self.dumpTable(t, level)
 		if (type(value) ~= 'function') then
 			if (type(value) == 'table') then
 				self.print(level .. attr .. ':')
-				self.dumpTable(value, level .. '    ')
+				self.dumpTable(value, level .. '	')
 			else
 				self.print(level .. attr .. ': ' .. tostring(value))
 			end