+++ dzVents/runtime/device-adapters/generic_device.lua
@@ -104,7 +104,7 @@ return {
 			device['lastUpdate'] = Time(data.lastUpdate)
 			device['rawData'] = data.rawData
 			device['nValue'] = data.data._nValue
-
+			device['sValue'] = data.data._state
 			device['cancelQueuedCommands'] = function()
 				domoticz.sendCommand('Cancel', {
 					type = 'device',