+++ dzVents/documentation/history.md
@@ -1,3 +1,8 @@
+[2.4.21]
+- fixed wrong direction for open() and close() for some types of blinds
+- Add inTable function to domoticz.utils
+- Add sValue attribute to devices
+
 [2.4.20]
 - Add quietOn() and quietOff() method to switchType devices